Card: Static method
Card now has a static method, instead of a singleton, that holds info on
the Bank, and is called in the toString().

Also added string variable 'owner' to card (for the toString) so when a
card is added to an account it needs to be passed the name of the
account holder.